requirements.txt
django
djangorestframework
gunicorn
whitenoise
psycopg2-binary
requests
urllib3